feat: full ts migration
