--- src/common/repository.js
@@ -1,10 +1,0 @@
-import { database } from '../db/database.js';
-
-/**
- * Class representing Repository
- */
-export class Repository {
-  constructor() {
-    this.db = database;
-  }
-}
