+++ src/common/repository.ts
@@ -1,0 +1,9 @@
+import { database, Database } from '../db/database';
+
+export class Repository {
+  protected db: Database;
+
+  constructor() {
+    this.db = database;
+  }
+}
