+++ src/middlewares/error.ts
@@ -1,0 +1,21 @@
+import { Response } from 'express';
+
+export class CustomError extends Error {
+  public statusCode: number;
+  public message: string;
+
+  constructor(statusCode: number, message: string) {
+    super();
+    this.statusCode = statusCode;
+    this.message = message;
+  }
+}
+
+export const handleError = (err: CustomError, res: Response): void => {
+  const { statusCode, message } = err;
+  res.status(statusCode).json({
+    status: 'error',
+    statusCode,
+    message,
+  });
+};
