+++ src/resources/boards/board.dto.ts
@@ -1,0 +1,9 @@
+export interface ColumnDto {
+  title: string;
+  order: number;
+}
+
+export interface BoardDto {
+  title: string;
+  columns: ColumnDto[];
+}
