+++ src/resources/boards/boards.controller.ts
@@ -1,50 +1,69 @@
-import { Router } from 'express';
-import { BoardsService } from './boards.service.js';
-import { Message } from '../../common/const.js';
+import { Router, Response, Request, NextFunction } from 'express';
+import { BoardsService } from './boards.service';
+import { Message } from '../../common/const';
+
+// const router = Router();
+// router.route('/').get(async (req, res, next) => {
+//
+// })
 
 export class BoardsController {
+  private boardsService: BoardsService;
+  public router: Router;
   constructor() {
     this.boardsService = new BoardsService();
     this.router = Router();
     this.routes();
   }
 
-  getAll = async (req, res) => {
+  getAll = async (_req: Request, res: Response): Promise<void> => {
     const boards = await this.boardsService.getAll();
     res.json(boards);
   };
 
-  create = async (req, res) => {
+  create = async (req: Request, res: Response): Promise<void> => {
     const boardDto = req.body;
     const board = await this.boardsService.create(boardDto);
     res.status(201).json(board);
   };
 
-  getById = async (req, res, next) => {
+  getById = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> => {
     const { id } = req.params;
     try {
-      const board = await this.boardsService.getById(id);
+      const board = await this.boardsService.getById(String(id));
       res.status(200).json(board);
     } catch (err) {
       next(err);
     }
   };
 
-  update = async (req, res, next) => {
+  update = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> => {
     const { id } = req.params;
     const boardDto = req.body;
     try {
-      const board = await this.boardsService.update(id, boardDto);
+      const board = await this.boardsService.update(String(id), boardDto);
       res.status(200).json(board);
     } catch (err) {
       next(err);
     }
   };
 
-  delete = async (req, res, next) => {
+  delete = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> => {
     const { id } = req.params;
     try {
-      await this.boardsService.deleteBoard(id);
+      await this.boardsService.deleteBoard(String(id));
       res.status(200).json({
         status: 'success',
         statusCode: res.statusCode,
@@ -55,7 +74,7 @@ export class BoardsController {
     }
   };
 
-  routes() {
+  routes(): void {
     this.router.get('/', this.getAll);
     this.router.post('/', this.create);
     this.router.get('/:id', this.getById);
