+++ src/resources/tasks/task.dto.ts
@@ -1,0 +1,8 @@
+export interface TaskDto {
+  title: string;
+  order: number;
+  description: string;
+  userId: string | null;
+  boardId: string;
+  columnId: string;
+}
