+++ src/resources/users/user.dto.ts
@@ -1,0 +1,11 @@
+export interface UserDto {
+  name: string;
+  login: string;
+  password: string;
+}
+
+export interface UserResponse {
+  id: string;
+  name: string;
+  login: string;
+}
