+++ src/resources/users/user.model.ts
@@ -1,29 +1,24 @@
 import { v4 as uuid } from 'uuid';
+import { UserDto, UserResponse } from './user.dto';
 
-/**
- * Class representing a user
- */
 export class User {
-  /**
-   *
-   * @param {string} id - user id
-   * @param {string} name - user name
-   * @param {string} login - user login
-   * @param {string} password - user password
-   */
+  public id;
+  public name;
+  public login;
+  public password;
+
   constructor({
-    id = uuid(),
     name = 'USER',
     login = 'user',
     password = 'P@55w0rd',
-  } = {}) {
-    this.id = id;
+  }: UserDto) {
+    this.id = uuid();
     this.name = name;
     this.login = login;
     this.password = password;
   }
 
-  static toResponse(user) {
+  static toResponse(user: User): UserResponse {
     const { id, name, login } = user;
     return { id, name, login };
   }
