+++ src/resources/users/users.controller.ts
@@ -1,51 +1,67 @@
-import { Router } from 'express';
-import { UsersService } from './users.service.js';
-import { User } from './user.model.js';
-import { Message } from '../../common/const.js';
+import { Router, Response, Request, NextFunction } from 'express';
+import { UsersService } from './users.service';
+import { User } from './user.model';
+import { Message } from '../../common/const';
+import { UserDto } from './user.dto';
 
 export class UsersController {
+  private userService: UsersService;
+  public router: Router;
+
   constructor() {
     this.userService = new UsersService();
     this.router = Router();
     this.routes();
   }
 
-  getAll = async (req, res) => {
+  getAll = async (_req: Request, res: Response): Promise<void> => {
     const users = await this.userService.getAll();
     res.json(users.map(User.toResponse));
   };
 
-  create = async (req, res) => {
-    const userDto = req.body;
+  create = async (req: Request, res: Response): Promise<void> => {
+    const userDto: UserDto = req.body;
     const user = await this.userService.create(userDto);
     res.status(201).json(User.toResponse(user));
   };
 
-  getById = async (req, res, next) => {
+  getById = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> => {
     const { id } = req.params;
     try {
-      const user = await this.userService.getById(id);
+      const user = await this.userService.getById(String(id));
       res.status(200).json(User.toResponse(user));
     } catch (err) {
       next(err);
     }
   };
 
-  update = async (req, res, next) => {
+  update = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> => {
     const { id } = req.params;
     const userDto = req.body;
     try {
-      const user = await this.userService.update(id, userDto);
+      const user = await this.userService.update(String(id), userDto);
       res.status(200).json(User.toResponse(user));
     } catch (err) {
       next(err);
     }
   };
 
-  delete = async (req, res, next) => {
+  delete = async (
+    req: Request,
+    res: Response,
+    next: NextFunction
+  ): Promise<void> => {
     const { id } = req.params;
     try {
-      await this.userService.deleteUser(id);
+      await this.userService.deleteUser(String(id));
       res.status(200).json({
         status: 'success',
         statusCode: res.statusCode,
@@ -56,7 +72,7 @@ export class UsersController {
     }
   };
 
-  routes() {
+  routes(): void {
     this.router.get('/', this.getAll);
     this.router.post('/', this.create);
     this.router.get('/:id', this.getById);
